extend server-side preprocessing for task specs/factor common task code out to their own file
