+++ server/task_specs/feature-apply.yml
@@ -3,7 +3,8 @@
 name: feature-apply
 mode: python
 
-script: "@include(feature-apply.py)"
+script: >
+  @include<feature-apply.py>
 
 inputs:
   - name: input_path_1
