+++ server/task_specs/feature-extract.yml
@@ -3,7 +3,8 @@
 name: feature-extract
 mode: python
 
-script: "@include(feature-extract.py)"
+script: >
+  @include<feature-extract.py>
 
 inputs:
   - name: input_path_1
