+++ server/task_specs/feature-match.yml
@@ -3,7 +3,8 @@
 name: feature-match
 mode: python
 
-script: "@include(feature-match.py)"
+script: >
+  @include<feature-match.py>
 
 inputs:
   - name: input_path_1
