+++ server/task_specs/iGPSe.yml
@@ -3,7 +3,8 @@
 name: iGPSe
 mode: r
 
-script: "@include(iGPSe.r)"
+script: >
+  @include<iGPSe.r>
 
 inputs:
   - name: mrna_input_path
