+++ server/task_specs/iGPSe2.yml
@@ -3,7 +3,8 @@
 name: iGPSePart2
 mode: r
 
-script: "@include(iGPSe2.r)"
+script: >
+  @include<iGPSe2.r>
 
 inputs:
   - name: transferData
