+++ server/task_specs/surv.yml
@@ -3,7 +3,8 @@
 name: Survival Plot
 mode: r
 
-script: "@include(surv.r)"
+script: >
+  @include<surv.r>
 
 inputs:
   - name: input_rdata
